Move tests to BaseConnectorSmokeTest
